+++ src/Testing/TestCase.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Butler\Service\Testing;
 
 use Illuminate\Foundation\Testing\TestCase as BaseTestCase;