+++ tests/Feature/AboutTest.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Butler\Service\Tests\Feature;
 
 use Butler\Service\Tests\TestCase;