+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.22.2] - 2023-05-16
+
 ### Added
 - A way to append items to navbar component.
 - Strict type declaration everywhere.